+++ src/DependencyInjection/SomnambulistApiExtension.php
@@ -9,8 +9,9 @@ use Somnambulist\Bundles\ApiBundle\Subscribers\ConvertJSONToPOSTRequestSubscribe
 use Somnambulist\Bundles\ApiBundle\Subscribers\RequestIdInjectorSubscriber;
 use Symfony\Component\Config\FileLocator;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
+use Symfony\Component\DependencyInjection\Extension\Extension;
 use Symfony\Component\DependencyInjection\Loader\XmlFileLoader;
-use Symfony\Component\HttpKernel\DependencyInjection\Extension;
+
 
 class SomnambulistApiExtension extends Extension
 {