+++ src/Request/Filters/Decoders/Behaviours/ConvertStringToArray.php
@@ -2,6 +2,7 @@
 
 namespace Somnambulist\Bundles\ApiBundle\Request\Filters\Decoders\Behaviours;
 
+use function json_validate;
 use function str_contains;
 use function str_getcsv;
 
@@ -9,7 +10,7 @@ trait ConvertStringToArray
 {
     protected function shouldBeArray(string $value, string $separator = ','): bool
     {
-        return str_contains($value, $separator);
+        return str_contains($value, $separator) && !json_validate($value);
     }
 
     protected function convertToArray(string $value, string $separator = ','): array