+++ tests/Request/Filters/Decoders/SimpleApiFilterDecoderTest.php
@@ -71,4 +71,67 @@ class SimpleApiFilterDecoderTest extends TestCase
             $this->assertEquals('=', $part->operator);
         }
     }
+
+    public function testApiDecoderHandlesJsonStrings()
+    {
+        $qb = new QueryBuilder();
+        $qb
+            ->where($qb->expr()->eq('this', 'that'))
+            ->andWhere($qb->expr()->eq('foo', 'bar'))
+            ->andWhere($qb->expr()->eq('baz', '[-34,-56,4,1,90,-11]'))
+            ->andWhere($qb->expr()->eq('bar', '{"bar":3,"baz":"four"}'))
+        ;
+
+        $queryString = http_build_query((new SimpleEncoder())->useNameForFiltersField('filters')->encode($qb));
+
+        $GET = [];
+        parse_str($queryString, $GET);
+
+        $formRequest = new SearchFormRequest(new Request($GET));
+        $parser      = new SimpleApiFilterDecoder();
+        $result      = $parser->useFiltersQueryName('filters')->decode($formRequest);
+
+        $this->assertCount(4, $result);
+
+        foreach ($result->parts() as $part) {
+            $this->assertInstanceOf(Expression::class, $part);
+            $this->assertEquals('=', $part->operator);
+
+            if (in_array($part->field, ['baz', 'bar'])) {
+                $this->assertIsString($part->value);
+            }
+        }
+    }
+
+    public function testApiDecoderHandlesCsvStrings()
+    {
+        $qb = new QueryBuilder();
+        $qb
+            ->where($qb->expr()->eq('this', 'that'))
+            ->andWhere($qb->expr()->eq('foo', 'bar'))
+            ->andWhere($qb->expr()->eq('baz', '-34,-56,4,1,90,-11'))
+        ;
+
+        $queryString = http_build_query((new SimpleEncoder())->useNameForFiltersField('filters')->encode($qb));
+
+        $GET = [];
+        parse_str($queryString, $GET);
+
+        $formRequest = new SearchFormRequest(new Request($GET));
+        $parser      = new SimpleApiFilterDecoder();
+        $result      = $parser->useFiltersQueryName('filters')->decode($formRequest);
+
+        $this->assertCount(3, $result);
+
+        foreach ($result->parts() as $part) {
+            $this->assertInstanceOf(Expression::class, $part);
+
+            if (in_array($part->field, ['baz'])) {
+                $this->assertEquals('IN', $part->operator);
+                $this->assertIsArray($part->value);
+            } else {
+                $this->assertEquals('=', $part->operator);
+            }
+        }
+    }
 }